Refactor GraphQL import paths to use lowercase filenames
