+++ app/routes/($locale).account.orders.$id.tsx
@@ -4,7 +4,7 @@ import { type LoaderFunctionArgs, json, redirect } from "@shopify/remix-oxygen";
 import clsx from "clsx";
 import invariant from "tiny-invariant";
 import type { OrderFragment } from "customer-accountapi.generated";
-import { CUSTOMER_ORDER_QUERY } from "~/graphql/customer-account/CustomerOrderQuery";
+import { CUSTOMER_ORDER_QUERY } from "~/graphql/customer-account/customer-order-query";
 import { statusMessage } from "~/lib/utils";
 import { Heading, PageHeader, Text } from "~/modules/text";
 import { Link } from "~/components/link";
